requirements.txt
numpy>=1.24
pandas>=2.0
pyarrow>=12.0
torch>=2.1
stable-baselines3>=2.1
gymnasium>=0.29
scikit-learn>=1.3
pyzmq>=25.0
MetaTrader5>=5.0.45
pytest>=7.4
matplotlib>=3.7
streamlit>=1.35
plotly>=5.20
python-telegram-bot>=21.0
